Deleted Question class
