+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
 
         <activity android:name=".activities.LoginActivity"/>
         <activity android:name=".activities.ValidateDiarrhoeaActivity">
-        <intent-filter>
+            <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
